+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/InstituteImpl.java
@@ -34,6 +34,8 @@ public class InstituteImpl implements Institute {
   @Column(nullable = false)
   private String alias;
   
+  private boolean archived;
+
   @ManyToOne(targetEntity = UserImpl.class, fetch = FetchType.LAZY)
   @JoinColumn(name = "createdBy", nullable = false)
   private User createdBy;
@@ -68,6 +70,16 @@ public class InstituteImpl implements Institute {
   @Override
   public void setAlias(String alias) {
     this.alias = alias;
+  }
+
+  @Override
+  public boolean isArchived() {
+    return archived;
+  }
+
+  @Override
+  public void setArchived(boolean archived) {
+    this.archived = archived;
   }
 
   @Override