+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultInstituteService.java
@@ -55,6 +55,7 @@ public class DefaultInstituteService implements InstituteService {
     authorizationManager.throwIfNonAdmin();
     Institute updatedInstitute = get(institute.getId());
     updatedInstitute.setAlias(institute.getAlias());
+    updatedInstitute.setArchived(institute.isArchived());
     User user = authorizationManager.getCurrentUser();
     updatedInstitute.setUpdatedBy(user);
     instituteDao.update(updatedInstitute);